Split FakeModelFactory from tests
