+++ core/helpers.py
@@ -1,7 +1,14 @@
 from django.http import JsonResponse
-from rest_framework import status
 from typing import Sequence, Iterable, Callable
+from rest_framework.response import Response
+from rest_framework import status
+from django.utils import timezone
 
+CURRENT_TZ = timezone.get_current_timezone()
+
+# --------------------------------------------------------------------------
+# 		Data Structures
+# --------------------------------------------------------------------------
 
 def filter_dict_keys(obj: dict, whitelist: Sequence):
 	return { k: v for k, v in obj.items() if k in whitelist }
@@ -23,6 +30,14 @@ def errorResponse(message, errors=tuple(), httpStatus=status.HTTP_400_BAD_REQUES
 	}
 	return JsonResponse(resp, status=httpStatus)
 
+def format_date(date) -> 'datetime':
+	"""
+	Format a date with the proper timezone
+	"""
+	if timezone.is_aware(date):
+		return date
+	else:
+		return timezone.make_aware(date, CURRENT_TZ, is_dst=False)
 
 def custom_editable_fields(request, obj=None, edition_readonly_fields=tuple(), always_readonly_fields=tuple()):
 	"""
